requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Google Cloud
google-cloud-storage>=2.14.0
google-api-python-client>=2.108.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-genai>=0.2.0
google-cloud-aiplatform>=1.38.0

# PDF processing
pdf2image>=1.17.0
Pillow>=10.1.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1

# Retry logic
tenacity>=8.2.3

# Logging
python-json-logger>=2.0.7

